refactor: repackaging the theme and renaming the components
