+++ design_system/src/main/java/com/baghdad/design_system/theme/Theme.kt
@@ -2,10 +2,10 @@ package com.baghdad.design_system.theme
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ReadOnlyComposable
-import com.baghdad.design_system.color.model.NovixColor
-import com.baghdad.design_system.color.model.localNovixColor
-import com.baghdad.design_system.textStyle.model.NovixTextStyle
-import com.baghdad.design_system.textStyle.model.localNovixTextStyle
+import com.baghdad.design_system.color.NovixColor
+import com.baghdad.design_system.color.localNovixColor
+import com.baghdad.design_system.textStyle.NovixTextStyle
+import com.baghdad.design_system.textStyle.localNovixTextStyle
 
 object Theme {
     val color: NovixColor
